+++ sql/query_04_focus_quality_checks.sql
@@ -3,5 +3,5 @@ SELECT
     count(*) FILTER (WHERE "BillingCurrency" !~ '^[A-Z]{3}$') AS invalid_currency_rows,
     count(*) FILTER (WHERE "ChargePeriodEnd" <= "ChargePeriodStart") AS invalid_charge_period_rows,
     count(*) FILTER (WHERE "EffectiveCost" IS NULL) AS missing_effective_cost_rows,
-    count(*) FILTER (WHERE "ProviderName" IS NULL OR "ServiceName" IS NULL) AS missing_provider_or_service_rows
+    count(*) FILTER (WHERE "ServiceProviderName" IS NULL OR "ServiceName" IS NULL) AS missing_provider_or_service_rows
 FROM focus.cost_usage;